+++ CHANGES.md
@@ -21,10 +21,17 @@ Changes:
 -
 
 Fixes:
-- fix recap.email docket_number parsing inside nested tags
+-
 
 
 ## Current
+
+**2.6.98 - 2025-12-05**
+
+Fixes:
+- fix recap.email docket_number parsing inside nested tags
+
+## Past
 
 **2.6.97 - 2025-11-28**
 
@@ -32,8 +39,6 @@ Fixes:
 - fix `alaska` scraper to skip header rows and prevent IndexError in table parsing #1674
 - fix `ky` and `kyctapp` IndexError when API returns empty document details #1666
 - fix `nyappdiv` and `nytrial` docket number extraction regex to require at least 1 number #1655
-
-## Past
 
 **2.6.96 - 2025-11-21**